+++ NEWS.md
@@ -1,3 +1,10 @@
+# antigen.garnish 0.0.5
+
+* add ncbi-blast functionality for near matches
+* add ncbi-blast functionality for known IEDB matches
+* implement fitness model from Lukza et al. **Nature** 2017 
+
+
 # antigen.garnish 0.0.4
 
 * Summary plots